change rationalnumber to double, finish line intersect

+++ src/Reader.cpp
@@ -14,23 +14,23 @@ void Reader::read() {
 		char type;
 		reader >> type;
 		if (type == 'L') {
-			int x1, y1, x2, y2;
+			long long x1, y1, x2, y2;
 			reader >> x1 >> y1 >> x2 >> y2;
 			g->addLine(DOUBLE_INFINITE_LINE, x1, x2, y1, y2);
 			//lines.push_back(Line(x1, y1, x2, y2));
 		}
 		else if (type == 'R') {
-			int x1, y1, x2, y2;
+			long long x1, y1, x2, y2;
 			reader >> x1 >> y1 >> x2 >> y2;
 			g->addLine(SINGLE_INFINITE_LINE, x1, x2, y1, y2);
 		}
 		else if (type == 'S') {
-			int x1, y1, x2, y2;
+			long long x1, y1, x2, y2;
 			reader >> x1 >> y1 >> x2 >> y2;
 			g->addLine(LIMITED_LINE, x1, x2, y1, y2);
 		}
 		else if (type == 'C') {
-			int a, b, r;
+			long long a, b, r;
 			reader >> a >> b >> r;
 			g->addCircle(a, b, r);
 			//circles.push_back(Circle(a, b, r));
